Move reusable fixtures to umi parent directory
